Added average rating to level info Changed average rating to float

+++ app/src/main/java/antitelegram/devenirchef/RateOthersActivity.java
@@ -22,8 +22,6 @@ import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.StorageReference;
 
-import org.w3c.dom.Text;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -150,7 +148,7 @@ public class RateOthersActivity extends DrawerBaseActivity {
         dataSnapshot.getRef().child(Constants.FINISHED_RECIPES).child(recipe.getIndex()).setValue(recipe);
 
         long newExp = dataSnapshot.child(Constants.EXP).getValue(Long.class) +
-            recipe.getAverageRating() * Constants.RATING_MULTIPLIER;
+            (long) recipe.getAverageRating() * Constants.RATING_MULTIPLIER;
 
         dataSnapshot.getRef().child(Constants.EXP).setValue(newExp);
